src/lambdas/api_auth/requirements.txt
python-jose[cryptography]==3.1.0
requests
